==> IMC/05a_IMC_secondary_clustering_3pix2.Rmd <==
---
title: "04_IMC_secondary_clustering"
author: "docaspar"
date: "10/25/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.height = 9, fig.width = 16, message = FALSE)
```

# Load libraries
```{r libs}
library(SingleCellExperiment)
library(tidyverse)
library(CATALYST)
library(scater)
library(rstudioapi)
library(readxl)
library(xlsx)
#library(clustree) Archive
library(Rphenograph)
library(clustree)
library(pals)
library(MetBrewer)
library(rasterpdf)
```

# Load SCE object
New SCE objects are created with 01_Generate_SCE_object.rmd for IMC data. 
```{r load_sce}
rm(list = ls())

setwd(dirname(getActiveDocumentContext()$path))

path.data = "../Data/cpout_3pix2/"
path.files <- "../R_Files/cpout_3pix2/"
path.out <- "../R_out/cpout_3pix2/"


ct <- c("Tregs", "memory CD4 T cells", "memory CD8 T cells", "CD8 Temra", "Macrophages", "cDCs", "Mast cells", "Fibroblasts", "Endothelium", "Epidermis", "Smooth muscle cells")
colvec_ct <- met.brewer("Signac", 19)[c(1, 4:6, 8, 10, 13, 16:19)]
names(colvec_ct) <- ct

sce = readRDS(file = paste0(path.files, "sce_3pix2_clustered_merged.rds"))
```

# T cell subclustering
## Exploration
```{r Tcells Phenograph, message=FALSE}
sceTcells <- sce[,colData(sce)$primary_clustering == "Tcells"] #This line replace filterSCE, which was buged somehow..
metadata(sceTcells) <- list()
coldata <- as.data.frame(colData(sceTcells))
colData(sceTcells) <- DataFrame(dplyr::select(coldata, - c("meta27")))

#load marker_classes for secondary clustering
secondary <- read_excel(paste0(path.files, "marker_classes_secondary_Tcells.xlsx"))
rowdata <- as.data.frame(rowData(sceTcells))
rowdata <- rowdata %>% dplyr::select(- c("marker_class", "used_for_clustering")) #replace primary marker classes with secondary
rowdata <- merge(rowdata, secondary, by = "Target")
rowData(sceTcells) <- DataFrame(rowdata[order(rowdata$channel),])
type_markers(sceTcells)

set.seed(42)
sceTcells <- runDR(sceTcells, dr = "UMAP", features = "type")
plotDR(sceTcells, "UMAP", color_by = type_markers(sceTcells), ncol = round(sqrt(16*length(type_markers(sceTcells))/9)), assay = "exprs")+
  geom_point(size = min(3, 5000/ncol(sceTcells)))+
  theme_cowplot() + 
  theme(plot.title = element_text(hjust = 0.5),
        strip.background = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank())
```

## T cell subclustering

```{r Tcells Phenograph, message=FALSE}
set.seed(42)
phenoresults <- Rphenograph(t(assay(filterSCE(sceTcells, marker_class == "type"), "exprs")), k = 15)

#Add phenograph cluster_ids to sce object
colData(sceTcells)$pheno <- as.factor(membership(phenoresults[[2]]))
sceTcells$cluster_id <- as.factor(membership(phenoresults[[2]]))
metadata(sceTcells)$cluster_codes <- data.frame(
    custom = factor(levels(as.factor(membership(phenoresults[[2]]))), levels = levels(as.factor(membership(phenoresults[[2]])))))

raster_pdf(paste0(path.out, "05a_3pix2_secondary_Tcells_phenograph.pdf"), height = 9, width = 16, res = 200)
plotExprHeatmap(sceTcells, features = "type",
    by = "cluster_id", k = "custom",
    scale = "first", q = 0.01, perc = TRUE, col_dend = FALSE, bars = TRUE, col_clust = FALSE)
dev.off()

raster_pdf(paste0(path.out, "05b_3pix2_secondary_Tcells_phenograph.pdf"), height = 10, width = 12, res = 200)
plotDR(sceTcells, "UMAP", color_by = "pheno") + geom_point(size = 2) + 
  scale_color_manual(values = brewer.paired(length(levels(as.factor(membership(phenoresults[[2]]))))))+ 
  geom_point(size = min(3, 5000/ncol(sceTcells)))+
  theme_cowplot()
dev.off()

raster_pdf(paste0(path.out, "05c_3pix2_secondary_Tcells_phenograph.pdf"), height = 4, width = 9, res = 200)
plotDR(sceTcells, "UMAP", color_by = type_markers(sceTcells), ncol = 4, assay = "exprs")+
  geom_point(size = min(3, 5000/ncol(sceTcells)))+
  theme_cowplot() + 
  theme(plot.title = element_text(hjust = 0.5),
        strip.background = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank())
dev.off()
```

## merge Tcell subclusters
```{r merge Tcells}
(merging_table1 <- read_excel(paste0(path.files, "merged_clusters_Tcells.xlsx")))
sceTcells <- mergeClusters(sceTcells, k = "custom", table = merging_table1, id = "merging1")

raster_pdf(paste0(path.out, "06a_3pix2_secondary_Tcells_merged.pdf"), height = 10, width = 12, res = 200)
plotDR(sceTcells, "UMAP", color_by = "merging1") + 
  geom_point(size = min(3, 5000/ncol(sceTcells)))+
  labs(title = "IMC T cells Secondary Clustering", col = "Cell type") +
  theme_cowplot() + 
  theme(plot.title = element_text(hjust = 0.5)) + 
  scale_color_manual(values = colvec_ct[1:4])
dev.off()

raster_pdf(paste0(path.out, "06b_3pix2_secondary_Tcells_merged.pdf"), height = 5, width = 12, res = 200)
plotDR(sceTcells, "UMAP", color_by = "merging1", facet_by = "Timepoint") + 
  geom_point(size = min(3, 5000/ncol(sceTcells)))+
  labs(title = "IMC T cells Secondary Clustering", col = "Cell type") + 
  scale_color_manual(values = colvec_ct[1:4])+
  theme_cowplot() + 
  theme(plot.title = element_text(hjust = 0.5),
        strip.background = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank())
dev.off()

code <- cluster_codes(sceTcells)[,c("custom", "merging1")]
coldata <- as.data.frame(colData(sceTcells))
coldata <- merge(coldata, code, by.x = "cluster_id", by.y = "custom")
colData(sceTcells) <- DataFrame(coldata[order(coldata$ImageNumber, coldata$CellNumber),])
```

# APC subclustering
## Exploration
```{r secondary clustering APCs, warning=FALSE, message=FALSE}
#Filter for cluster/cell type
sceAPCs <- sce[,colData(sce)$primary_clustering == "APCs"]
metadata(sceAPCs) <- list()
coldata <- as.data.frame(colData(sceAPCs))
colData(sceAPCs) <- DataFrame(coldata)

#load marker_classes for secondary clustering
secondary <- read_excel(paste0(path.files, "marker_classes_secondary_APCs.xlsx"))
rowdata <- as.data.frame(rowData(sceAPCs))
rowdata <- rowdata %>% dplyr::select(- c("marker_class", "used_for_clustering")) #replace primary marker classes with secondary
rowdata <- merge(rowdata, secondary, by = "Target")
rowData(sceAPCs) <- DataFrame(rowdata[order(rowdata$channel),])


set.seed(1234)
sceAPCs <- runDR(sceAPCs, dr = "UMAP", features = "type")
```

## FlowSOM clustering
```{r secondary clustering APCs, warning=FALSE, message=FALSE}
#secondary clustering

ksec <- 14
set.seed(1234)
sceAPCs <- cluster(sceAPCs, features = type_markers(sceAPCs),
    xdim = 10, ydim = 10, maxK = ksec, seed = 1234)

raster_pdf(paste0(path.out, "07a_3pix2_secondary_APCs_flowSOM.pdf"), height = 10, width = 12, res = 200)
plotDR(sceAPCs, "UMAP", color_by = paste0("meta", ksec)) + geom_point(size = (5000/ncol(sceAPCs)))+
  theme_cowplot()
dev.off()

raster_pdf(paste0(path.out, "07b_3pix2_secondary_APCs_flowSOM.pdf"), height = 4, width = 9, res = 200)
plotDR(sceAPCs, "UMAP", color_by = type_markers(sceAPCs), ncol = 4, assay = "exprs")+ 
  geom_point(size = min(3, 5000/ncol(sceAPCs)))+
  theme_cowplot() + 
  theme(strip.background = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank())
dev.off()

pdf(paste0(path.out, "07c_3pix2_secondary_APCs_flowSOM.pdf"), height = 9, width = 16, onefile = TRUE)
plotExprHeatmap(sceAPCs, features = "type",
    by = "cluster_id", k = paste0("meta", ksec),
    scale = "first", q = 0.01, perc = TRUE, col_dend = FALSE, bars = TRUE, col_clust = FALSE)

clustree(cluster_codes(sceAPCs), prefix = "meta")
dev.off()
```


```{r secondary merging APCs, warning=FALSE, message=FALSE}
merging_table1 <- read_excel(paste0(path.files, "merged_clusters_APCs_temp_DC.xlsx")) #This merging table combines alls cDC subsets to cDCs
#merging_table1 <- read_excel(paste0(path.files, "merged_clusters_APCs_temp.xlsx")) #This merging table keeps all cDC subsets separate

sceAPCs <- mergeClusters(sceAPCs, k = "meta14", table = merging_table1, id = "merging1")
code <- cluster_codes(sceAPCs)[,c("som100", "merging1")]
coldata <- as.data.frame(colData(sceAPCs))
coldata <- merge(coldata, code, by.x = "cluster_id", by.y = "som100")
coldata$merging1 <- factor(coldata$merging1, levels = levels(coldata$merging1))
#coldata[coldata$pheno == 14, "merging1"] <- "Mast cells"

colData(sceAPCs) <- DataFrame(coldata[order(coldata$ImageNumber, coldata$CellNumber),])

raster_pdf(paste0(path.out, "08a_3pix2_secondary_APCs_merged.pdf"), height = 10, width = 12, res = 200)
plotDR(sceAPCs, "UMAP", color_by = "merging1", assay = "exprs")+ 
    scale_color_manual(values = colvec_ct[5:7])+
    geom_point(size = min(3, 5000/ncol(sceAPCs)))+
  theme_cowplot()
dev.off()

raster_pdf(paste0(path.out, "08b_3pix2_secondary_APCs_merged.pdf"), height = 5, width = 12, res = 200)
plotDR(sceAPCs, "UMAP", color_by = "merging1", assay = "exprs") + 
    facet_wrap("Timepoint") + 
    scale_color_manual(values = colvec_ct[5:7])+
    geom_point(size = min(3, 5000/ncol(sceAPCs)))+
  theme_cowplot()
dev.off()

```

# merging subclusters into main SCE
```{r merge subclusters}
tcell_cluster_id <- as.data.frame(colData(sceTcells)[,c("cellID", "merging1")])
apc_cluster_id <- as.data.frame(colData(sceAPCs)[,c("cellID", "merging1")])
subcluster_id <- rbind(tcell_cluster_id, apc_cluster_id)
colnames(subcluster_id) <- c("cellID", "subclusters")
#table(duplicated(subcluster_id$cellID)) #make sure all cell IDs are unique

coldata <- data.frame(colData(sce)) %>% dplyr::select(- "meta27")
me <- merge(coldata, subcluster_id, by = c("cellID"), all = TRUE) 
me <- me[order(me$ImageNumber, me$CellNumber),]
me$sample_id <- factor(me$sample_id, levels = paste(1:17))
me$primary_clustering<- as.character(me$primary_clustering)
me$subclusters <- as.character((me$subclusters))
me[is.na(me$subclusters),"subclusters"] <- me[is.na(me$subclusters),"primary_clustering"]

me$cluster_id <- factor(me$subclusters)
colData(sce) <- DataFrame(me)

metadata(sce)$cluster_codes <- data.frame(
    custom = factor(levels(me$cluster_id), levels = levels(me$cluster_id)))
table(colData(sce)[,c("subclusters", "Timepoint")])

raster_pdf(paste0(path.out, "09a_3pix2_subclustered.pdf"), height = 10, width = 12, res = 200)
plotDR(sce, "UMAP", color_by = "subclusters", assay = "exprs") + 
  labs(title = "IMC APCs Secondary Clustering", col = "Cell type")+
  theme_cowplot() + 
  theme(plot.title = element_text(hjust = 0.5)) +  
  scale_color_manual(values = colvec_ct) + 
  geom_point(size = min(3, 5000/ncol(sce)))
dev.off()

raster_pdf(paste0(path.out, "09b_3pix2_subclustered.pdf"), height = 5, width = 12, res = 200)
plotDR(sce, "UMAP", color_by = "subclusters", assay = "exprs") + 
  labs(title = "IMC APCs Secondary Clustering", col = "Cell type")+
  theme_cowplot() + 
  theme(plot.title = element_text(hjust = 0.5)) + 
  scale_color_manual(values = colvec_ct) + 
  facet_wrap("Timepoint")+ 
  geom_point(size = min(3, 5000/ncol(sce)))
dev.off()
```
# Save SCE object
```{r save SCE}
saveRDS(sce,file = paste0(path.files, "sce_3pix2_subclustered.rds"))
```


# Session Info
```{r sessionInfo}
sessionInfo()
```

# Archive
### flowSOM Tcells
```{r secondary clustering Tcells, warning=FALSE, message=FALSE, fig.height=9, fig.width=16, eval=FALSE, include=FALSE}
# #Filter for cluster/cell type
# sceTcells <- filterSCE(sce, merging1 == "Tcells")
# metadata(sceTcells) <- list()
# coldata <- as.data.frame(colData(sceTcells))
# colData(sceTcells) <- DataFrame(dplyr::select(coldata, - c("meta27")))
# 
# #load marker_classes for secondary clustering
# secondary <- read_excel(paste0(path, "cpout_3pix2/marker_classes_secondary_Tcells.xlsx"))
# rowdata <- as.data.frame(rowData(sceTcells))
# rowdata <- rowdata %>% dplyr::select(- c("marker_class", "used_for_clustering")) #replace primary marker classes with secondary
# rowdata <- merge(rowdata, secondary, by = "Target")
# rowData(sceTcells) <- DataFrame(rowdata[order(rowdata$channel),])
# type_markers(sceTcells)
# 
# ksec <- 14
# set.seed(42)

#secondary clustering
# sceTcells <- cluster(sceTcells, features = type_markers(sceTcells),
#     xdim = 8, ydim = 8, maxK = ksec, seed = 1234)
# 
# sceTcells <- runDR(sceTcells, dr = "UMAP", features = "type")
# 
# #Create PDF with standard clustering output
# #pdf("~/Desktop/secondary_clustering_Tcells_flowSOM.pdf", height = 9, width = 16, onefile = TRUE)
# 
# clustree(cluster_codes(sceTcells), prefix = "meta")
# 
# plotExprHeatmap(sceTcells, features = "type",
#     by = "cluster_id", k = paste0("meta", ksec),
#     scale = "first", q = 0.01, perc = TRUE, col_dend = FALSE, bars = TRUE, col_clust = FALSE)
# 
# plotClusterExprs(sceTcells, k = paste0("meta", ksec), features = "type") 
# 
# plotDR(sceTcells, "UMAP", color_by = paste0("meta", ksec)) + geom_point(size = 2)#or a cluster, apparently -> names(cluster_codes(sce))
# 
# # # Facet per condition
# # plotDR(sceTcells, "UMAP", color_by = paste0("meta", ksec)) + facet_wrap("Timepoint") +
# #     guides(color = guide_legend(ncol = 2, override.aes = list(size = 3))) + geom_point(size = 2)
# 
# plotDR(sceTcells, "UMAP", color_by = type_markers(sceTcells), ncol = round(sqrt(16*length(type_markers(sceTcells))/9)), assay = "exprs")
# 
# #dev.off()

# (merging_table1 <- read_excel(paste0(path, "cpout_3pix2/merged_clusters_Tcells.xlsx")))
# sceTcells <- mergeClusters(sceTcells, k = "meta14", table = merging_table1, id = "merging1")
# plotDR(sceTcells, "UMAP", color_by = "merging1") + geom_point(size = 2)
# 
# code <- cluster_codes(sceTcells)[,c("som100", paste0("meta", ksec), "merging1")]
# #code <- cluster_codes(sceTcells)[,c("som100", paste0("meta", ksec))]
# #colnames(code) <- c("som100", "merging1") #just testing, remove
# coldata <- as.data.frame(colData(sceTcells))
# coldata <- merge(coldata, code, by.x = "cluster_id", by.y = "som100")
# colData(sceTcells) <- DataFrame(coldata[order(coldata$ImageNumber, coldata$CellNumber),])
# #colData(sceTcells)
# table(colData(sceTcells)[,c("merging1", "Timepoint")])
```



# Archive
# APC cluster
## secondary clustering APCs
```{r pheno APCs, message=FALSE, eval=FALSE, include=FALSE}
set.seed(1234)
#Filter for cluster/cell type
sceAPCs <- sce[,colData(sce)$primary_clustering == "APCs"]
metadata(sceAPCs) <- list()
coldata <- as.data.frame(colData(sceAPCs))
colData(sceAPCs) <- DataFrame(dplyr::select(coldata, - c("meta27")))

#load marker_classes for secondary clustering
(secondary <- read_excel(paste0(path, "marker_classes_secondary_APCs.xlsx")))
rowdata <- as.data.frame(rowData(sceAPCs))
rowdata <- rowdata %>% dplyr::select(- c("marker_class", "used_for_clustering")) #replace primary marker classes with secondary
rowdata <- merge(rowdata, secondary, by = "Target")
rowData(sceAPCs) <- DataFrame(rowdata[order(rowdata$channel),])
type_markers(sceAPCs)

phenoresults <- Rphenograph(t(assay(filterSCE(sceAPCs, marker_class == "type"), "exprs")), k = 20)

colData(sceAPCs)$pheno <- as.factor(membership(phenoresults[[2]]))
sceAPCs$cluster_id <- as.factor(membership(phenoresults[[2]]))
metadata(sceAPCs)$cluster_codes <- data.frame(
    custom = factor(levels(as.factor(membership(phenoresults[[2]]))), levels = levels(as.factor(membership(phenoresults[[2]])))))

plotDR(sceAPCs, "UMAP", color_by = "pheno") + geom_point(size = 2)
plotDR(sceAPCs, "UMAP", color_by = type_markers(sceAPCs), ncol = round(sqrt(16*length(type_markers(sceAPCs))/9)), assay = "exprs")
#dev.off()

table(colData(sceAPCs)[c("pheno", "Timepoint")])
colSums(table(colData(sceAPCs)[c("cluster_id", "Timepoint")]))
```


## merging secondary APC clusters
```{r merge clusters APCs, fig.height=9, fig.width=16, eval=FALSE, include=FALSE}
(merging_table1 <- read_excel(paste0(path, "merged_clusters_APCs_pheno.xlsx")))
sceAPCs <- mergeClusters(sceAPCs, k = "custom", table = merging_table1, id = "merging1")
#pdf(paste0(path, "IMC_R_output/merged_APCs_UMAP.pdf"), height = 9, width = 16, onefile = TRUE)
plotDR(sceAPCs, "UMAP", color_by = "merging1", assay = "exprs") + geom_point(size = 2) + labs(title = "IMC APCs Secondary Clustering", col = "Cell type") + theme(plot.title = element_text(hjust = 0.5)) + scale_color_brewer(palette = "Set1")
#dev.off()

code <- cluster_codes(sceAPCs)[,c("custom", "merging1")]
coldata <- as.data.frame(colData(sceAPCs))
coldata <- merge(coldata, code, by.x = "cluster_id", by.y = "custom")
colData(sceAPCs) <- DataFrame(coldata[order(coldata$ImageNumber, coldata$CellNumber),])
colData(sceAPCs) %>% data.frame() %>% head()

table(colData(sceAPCs)[,c("merging1", "Timepoint")])
```



```{r merge sub archive, eval=FALSE, include=FALSE}
# tcell_cluster_id <- as.data.frame(colData(sceTcells)[,c("cellID", "merging1")])
# apc_cluster_id <- as.data.frame(colData(sceAPCs)[,c("cellID", "merging1")])
# subcluster_id <- rbind(tcell_cluster_id, apc_cluster_id)
# colnames(subcluster_id) <- c("cellID", "subclusters")
# #table(duplicated(subcluster_id$cellID)) #make sure all cell IDs are unique
# 
# coldata <- as.data.frame(colData(sce)) %>% dplyr::select(-meta20)
# me <- merge(coldata, subcluster_id, by = c("cellID"), all = TRUE) 
# me <- me[order(me$ImageNumber, me$CellNumber),]
# #coldata[coldata$cellID %in% tcell_cluster_id$cellID,"subsclusters"] <- tcell_cluster_id$merging1
# 
# me$primary_clustering<- as.character(me$primary_clustering)
# me$subclusters <- as.character((me$subclusters))
# me
# me[is.na(me$subclusters),"subclusters"] <- me[is.na(me$subclusters),"primary_clustering"]
# colData(sce) <- DataFrame(me)
# table(colData(sce)$subclusters)
# saveRDS(sce,file = "../../C - IICT SLE IL-2/I - Imaging mass cytometry/SCEobjects/sce_2_3pix2_subclustered.rds")
# table(colData(sce)[,c("subclusters", "Timepoint")])
```
