==> IMC/11_Secondary_Cluster_Visualization_combined.Rmd <==
---
title: "06_secondary_cluster_visualization"
author: "docaspar"
date: "11/10/2021"
output: html_document
---

# Preparations

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE, root.dir = ".")
```

# Libraries

```{r}
library(SingleCellExperiment)
library(cytomapper)
library(stringr)
library(CATALYST)
library(readxl)
library(MetBrewer)
```

# Setup
```{r}
#Clear R's memory
rm(list=ls())

path.files <- "../R_Files/cpout_combined/"
path.out <- "../R_out/cpout_combined/"

# ct <- c("Tregs", "memory CD4 T cells", "memory CD8 T cells", "CD8 Temra", "Macrophages", "cDCs", "Mast cells", "Fibroblasts", "Endothelium", "Epidermis", "Smooth muscle cells")
# colvec_ct <- met.brewer("Signac", 19)[c(1, 4:6, 8, 10, 13, 16:19)]
# names(colvec_ct) <- ct

ct <- c("Tregs", "memory CD4 T cells", "memory CD8 T cells", "CD8 Temra", "Macrophages", "cDCs", "Mast cells", "Fibroblasts", "Endothelium", "Epidermis", "Smooth muscle cells")
colvec_ct <- met.brewer("Signac", 19)[c(6, 4, 3, 1, 8, 10, 13, 16:19)]
names(colvec_ct) <- ct
```


# Read the data, masks and pixel information (IMC images)

```{r load data}
sce <- readRDS(file = paste0(path.files, "sce_3pix2_batch2_subclustered_combined.rds"))

all_pixel <- readRDS(paste0(path.files, "all_pixel.rds"))
all_mask <- readRDS(paste0(path.files, "all_mask.rds"))
```


```{r project clusters on images, fig.width=16, fig.height=9}
pdf(paste0(path.out, "10_CellMasks_subclustered.pdf"), height = 9, width = 16, onefile = TRUE)
plotCells(mask = all_mask, object = sce,
            cell_id = "CellNumber", img_id = "ImageNumber", 
            colour_by = "subclusters", exprs_values = "exprs", margin = 3, display = "all", legend = list(colour_by.legend.cex = 0.5))
plotCells(mask = all_mask, object = sce,
            cell_id = "CellNumber", img_id = "ImageNumber", 
            colour_by = "subclusters", exprs_values = "exprs", margin = 3, display = "single", legend = list(colour_by.legend.cex = 0.75))
dev.off()
```


```{r, fig.height=9, fig.width=16}
pdf(paste0(path.out, "11d_Pixels_fig1f_newcol.pdf"), height = 9, width = 16, onefile = TRUE)
images <- c(19)
plotPixels(all_pixel[images,], mask = all_mask[images,], object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD3", "CD4", "CD25", "Ecadherin", "CollagenI", "CD163"),
            bcg = list(CD3 = c(0,30,1),
                       CD4 = c(0,20,1),
                       CD25 = c(0,20,1),
                       Ecadherin = c(0,2,1),
                       CollagenI = c(0,20,1),
                       CD163 = c(0,10,1)),
           colour = list(CD3 = c("black", "green"),
                      CD4 = c("black", "blue"),
                      CD25 = c("black", "red"),
                      Ecadherin = c("black", "yellow"),
                      CollagenI = c("black", "goldenrod3"),
                      CD163 = c("black", "orange"),
                      subclusters = colvec_ct),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(images))))
dev.off()

pdf(paste0(path.out, "11e_Pixels_fig1f_CD3CD4CD25EcadColI.pdf"), height = 9, width = 16, onefile = TRUE)
images <- c(19)
plotPixels(all_pixel[images,], mask = all_mask[images,], object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD3", "CD4", "CD25", "Ecadherin", "CollagenI"),
            bcg = list(CD3 = c(0,30,1),
                       CD4 = c(0,20,1),
                       CD25 = c(0,20,1),
                       Ecadherin = c(0,2,1),
                       CollagenI = c(0,20,1)),
           colour = list(CD3 = c("black", "green"),
                      CD4 = c("black", "blue"),
                      CD25 = c("black", "red"),
                      Ecadherin = c("black", "yellow"),
                      CollagenI = c("black", "goldenrod3"),
                      subclusters = colvec_ct),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(images))))
dev.off()


pdf(paste0(path.out, "11b_Pixels_fig1f.pdf"), height = 9, width = 16, onefile = TRUE)
images <- c(1:27)
plotPixels(all_pixel[images,], mask = all_mask[images,], object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD3", "CD4", "CD25", "Ecadherin", "CollagenI", "CD163"),
            bcg = list(CD3 = c(0,30,1),
                       CD4 = c(0,20,1),
                       CD25 = c(0,20,1),
                       Ecadherin = c(0,1,1),
                       #DNA1 = c(0,20,1),
                       CollagenI = c(0,20,1),
                       CD163 = c(0,10,1)),
           colour = list(CD3 = c("black", "red"),
                      CD4 = c("black", "blue"),
                      CD25 = c("black", "green"),
                      Ecadherin = c("black", "yellow"),
                      #DNA1 = c("black", "white"),
                      CollagenI = c("black", "goldenrod3"),
                      CD163 = c("black", "gray75"),
                      subclusters = colvec_ct),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(images))))
dev.off()
```


```{r, fig.height=9, fig.width=16}
pdf(paste0(path.out, "11c_Pixels_fig1f_no_background.pdf"), height = 9, width = 16, onefile = TRUE)
images <- c(19)
plotPixels(all_pixel[images,], mask = all_mask[images,], object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD3", "CD4", "CD25", "Ecadherin", "CD163"),
            bcg = list(CD3 = c(0,30,1),
                       CD4 = c(0,20,1),
                       CD25 = c(0,20,1),
                       Ecadherin = c(0,1,1),
                       #DNA1 = c(0,20,1),
                       CD163 = c(0,10,1)),
           colour = list(CD3 = c("black", "red"),
                      CD4 = c("black", "blue"),
                      CD25 = c("black", "green"),
                      Ecadherin = c("black", "yellow"),
                      #DNA1 = c("black", "white"),
                      CD163 = c("black", "goldenrod3"),
                      subclusters = colvec_ct),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(images))))
dev.off()
```


```{r}
pdf(paste0(path.out, "plotPixels_3pix2_sub_structural.pdf"), height = 9, width = 16, onefile = TRUE)
plotPixels(all_pixel, mask = all_mask, object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD45", "SMA", "CD141", "DNA1"),
            bcg = list(CD45 = c(0,40,1),
                       SMA = c(0,5,1),
                       CD141 = c(0,30,1),
                       DNA1 = c(0,20,1)),
           colour = list(CD45 = c("black", "red"),
                      SMA = c("black", "blue"),
                      CD141 = c("black", "green"),
                      DNA1 = c("black", "white")),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(seq(1:length(all_pixel))))),
           scale_bar = NULL)
dev.off()

pdf(paste0(path.out, "plotPixels_3pix2_sub_Treg.pdf"), height = 9, width = 16, onefile = TRUE)
plotPixels(all_pixel, mask = all_mask, object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD4", "CD25", "FOXp3", "CD27","DNA1"),
            bcg = list(CD4 = c(0,50,1),
                       CD25 = c(0,50,1),
                       FOXp3 = c(0,50,1),
                       CD27 = c(0,40,1),
                       DNA1 = c(0,10,1)),
           colour = list(CD4 = c("black", "red"),
                      CD25 = c("black", "blue"),
                      FOXp3 = c("black", "green"),
                      CD27 = c("black", "orange"),
                      DNA1 = c("black", "white")),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(seq(1:length(all_pixel))))),
           scale_bar = NULL)
dev.off()

pdf(paste0(path.out, "plotPixels_3pix2_sub_APC.pdf"), height = 9, width = 16, onefile = TRUE)
plotPixels(all_pixel, mask = all_mask, object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD45", "HLADR", "CD163", "DNA1"),
            bcg = list(CD45 = c(0,40,1),
                       HLADR = c(0,10,1),
                       CD163 = c(0,10,1),
                       DNA1 = c(0,10,1)),
           colour = list(CD45 = c("black", "red"),
                      HLADR = c("black", "blue"),
                      CD163 = c("black", "green"),
                      DNA1 = c("black", "white")),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(seq(1:length(all_pixel))))),
           scale_bar = NULL)
dev.off()
```


```{r, eval=FALSE}
colData(sce) %>% as.data.frame() %>% tail
sceTreg <- sce[,colData(sce)$subclusters == "Treg"]

pdf(paste0(path.out, "plotPixels_3pix2_sub_Treg_only.pdf"), height = 9, width = 16, onefile = TRUE)
plotPixels(all_pixel, mask = all_mask, object = sceTreg,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD25", "FOXp3", "CD27","DNA1"),
            bcg = list(CD25 = c(0,50,1),
                       FOXp3 = c(0,50,1),
                       CD27 = c(0,40,1),
                       DNA1 = c(0,10,1)),
           colour = list(CD25 = c("black", "blue"),
                      FOXp3 = c("black", "green"),
                      CD27 = c("black", "red"),
                      DNA1 = c("black", "white"),
                      subclusters = c(Treg = "purple")),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 0.75),
           image_title = list(text = c(paste(seq(1:length(all_pixel))))),
           scale_bar = NULL)
dev.off()
```



I believe this creates mismatched links between mask ids and cell ids because cells from the two batches were collected and images were not.
The first 17 images (batch 1) should be fine, but batch 2 will likely be faulty.
```{r, fig.height=9, fig.width=16}
all_pixel[1,]
pdf(paste0(path.out, "plotPixels_3pix2_sub_Tcell.pdf"), height = 9, width = 16, onefile = TRUE)
plotPixels(all_pixel[1,], mask = all_mask[1,], object = sce,
           img_id = "ImageNumber", cell_id = "CellNumber",
           colour_by = c("CD3", "CD4", "CD8", "DNA1"),
            bcg = list(CD3 = c(0,30,1),
                       CD4 = c(0,70,1),
                       CD8 = c(0,70,1),
                       DNA1 = c(0,20,1)),
           colour = list(CD3 = c("black", "red"),
                      CD4 = c("black", "blue"),
                      CD8 = c("black", "green"),
                      DNA1 = c("black", "white")),
           outline_by = "subclusters", display = "single", legend = list(colour_by.legend.cex = 0.75,
                                                                         colour_by.title.cex = 2),
           image_title = list(text = c(paste(seq(1:length(all_pixel[1,]))))))
dev.off()
```
